Fix review issues: overflow bug, input validation, assign guard, undo simplification
- Fix Candidates::all overflow when n=16 by computing in u32 before casting
- Add n range check (0 < n <= 16) and clue/board size consistency in SolveState::new
- Add early return in assign for already-assigned cells to prevent contradiction
- Store prev_grid in UndoEntry to simplify undo logic

+++ skyscrapers-solver/src/candidates.rs
@@ -8,7 +8,7 @@ impl Candidates {
     /// Creates a candidate set containing all values 1..=n.
     pub(crate) fn all(n: u8) -> Self {
         debug_assert!(n <= 16);
-        Self((1u16 << n) - 1)
+        Self(((1u32 << n) - 1) as u16)
     }
 
     /// Creates an empty candidate set.
